Changed setting of currentUser context.
In the lecture project a new user object was created for current user context which only contains the id and the mail of the current user. This means that I had to look for all additional information in the persistence context each time. For this reason I set the currentUser context as an attached entity by get it from userRepository.findById(id).